[Cloud Security] Vulnerabilities Preview & Refactor CSP Plugin PHASE 1 (#193385)
## Summary
In an attempt to make Reviewing easier and more accurate, the
implementation of Vulnerabilities on Host.name flyout in Alerts Page
will be split into 2 Phases

Phase 1: Move Functions, Utils or Helpers, Hooks, constants to Package
Phase 2: Implementing the feature

(cherry picked from commit 434430a995487971137ce6b3182d2285758cd4c1)
