+++ x-pack/packages/kbn-cloud-security-posture-common/constants.ts
@@ -26,3 +26,10 @@ export const CLOUD_SECURITY_POSTURE_BASE_PATH = '/cloud_security_posture';
 export const CDR_MISCONFIGURATIONS_DATA_VIEW_ID_PREFIX =
   'security_solution_cdr_latest_misconfigurations';
 export const SECURITY_DEFAULT_DATA_VIEW_ID = 'security-solution-default';
+
+export const CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN =
+  'logs-cloud_security_posture.vulnerabilities_latest-default';
+export const CDR_LATEST_THIRD_PARTY_VULNERABILITIES_INDEX_PATTERN =
+  'security_solution-*.vulnerability_latest';
+export const CDR_VULNERABILITIES_INDEX_PATTERN = `${CDR_LATEST_THIRD_PARTY_VULNERABILITIES_INDEX_PATTERN},${CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN}`;
+export const LATEST_VULNERABILITIES_RETENTION_POLICY = '3d';
