+++ x-pack/packages/kbn-cloud-security-posture-common/index.ts
@@ -19,9 +19,11 @@ export type {
 } from './types/status';
 export type { CspFinding, CspFindingResult } from './types/findings';
 export type { BenchmarksCisId } from './types/benchmark';
+export type { VulnSeverity } from './types/vulnerabilities';
 export * from './constants';
 export {
   extractErrorMessage,
   buildMutedRulesFilter,
   buildEntityFlyoutPreviewQuery,
 } from './utils/helpers';
+export { getAbbreviatedNumber } from './utils/get_abbreviated_number';
