+++ x-pack/packages/kbn-cloud-security-posture-common/schema/vulnerabilities/csp_vulnerability_finding.ts
@@ -7,7 +7,7 @@
 
 // TODO: this needs to be defined in a versioned schema
 import type { EcsEvent } from '@elastic/ecs';
-import { VulnSeverity } from '../types_old';
+import type { VulnSeverity } from '../../types/vulnerabilities';
 
 export interface CspVulnerabilityFinding {
   '@timestamp': string;
