+++ x-pack/packages/kbn-cloud-security-posture-common/schema/vulnerabilities/latest.ts
@@ -4,4 +4,5 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+
 export * from './csp_vulnerability_finding';
