+++ x-pack/packages/kbn-cloud-security-posture-common/utils/get_abbreviated_number.ts
