+++ x-pack/plugins/cloud_security_posture/public/common/utils/get_vulnerability_colors.ts
@@ -6,8 +6,8 @@
  */
 
 import { euiThemeVars } from '@kbn/ui-theme';
+import type { VulnSeverity } from '@kbn/cloud-security-posture-common';
 import { VULNERABILITIES_SEVERITY } from '../../../common/constants';
-import { VulnSeverity } from '../../../common/types_old';
 
 export const getCvsScoreColor = (score: number): string | undefined => {
   if (score <= 4) {
