+++ x-pack/plugins/cloud_security_posture/public/components/cloud_security_data_table/additional_controls.tsx
@@ -8,9 +8,9 @@ import React, { FC, PropsWithChildren } from 'react';
 import { EuiButtonEmpty, EuiFlexItem } from '@elastic/eui';
 import { type DataView } from '@kbn/data-views-plugin/common';
 import { FormattedMessage } from '@kbn/i18n-react';
+import { getAbbreviatedNumber } from '@kbn/cloud-security-posture-common';
 import { FieldsSelectorModal, useFieldsModal } from './fields_selector';
 import { useStyles } from './use_styles';
-import { getAbbreviatedNumber } from '../../common/utils/get_abbreviated_number';
 import { CSP_FIELDS_SELECTOR_OPEN_BUTTON } from '../test_subjects';
 
 const GroupSelectorWrapper: FC<PropsWithChildren<unknown>> = ({ children }) => {
