+++ x-pack/plugins/cloud_security_posture/public/components/vulnerability_badges.tsx
@@ -9,8 +9,8 @@ import { EuiBadge, EuiIcon, EuiTextColor } from '@elastic/eui';
 import React from 'react';
 import { css } from '@emotion/react';
 import { float } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
+import type { VulnSeverity } from '@kbn/cloud-security-posture-common';
 import { getCvsScoreColor, getSeverityStatusColor } from '../common/utils/get_vulnerability_colors';
-import { VulnSeverity } from '../../common/types_old';
 import { VULNERABILITIES_CVSS_SCORE_BADGE_SUBJ } from './test_subjects';
 
 interface CVSScoreBadgeProps {
