+++ x-pack/plugins/cloud_security_posture/public/components/vulnerability_severity_map.tsx
@@ -15,9 +15,9 @@ import {
   EuiText,
 } from '@elastic/eui';
 import { PaletteColorStop } from '@elastic/eui/src/components/color_picker/color_palette_picker';
+import type { VulnSeverity } from '@kbn/cloud-security-posture-common';
 import { i18n } from '@kbn/i18n';
 import { getSeverityStatusColor } from '../common/utils/get_vulnerability_colors';
-import { VulnSeverity } from '../../common/types_old';
 import { SeverityStatusBadge } from './vulnerability_badges';
 
 interface Props {
