+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/_mocks_/vulnerability.mock.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { CspVulnerabilityFinding } from '../../../../common/schemas';
+import type { CspVulnerabilityFinding } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 
 export const mockVulnerabilityHit: CspVulnerabilityFinding = {
   '@timestamp': '2023-03-30T10:27:35.013Z',
