+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/hooks/use_latest_vulnerabilities_grouping.tsx
@@ -14,6 +14,7 @@ import {
   parseGroupingQuery,
 } from '@kbn/grouping/src';
 import { useMemo } from 'react';
+import { LATEST_VULNERABILITIES_RETENTION_POLICY } from '@kbn/cloud-security-posture-common';
 import { buildEsQuery, Filter } from '@kbn/es-query';
 import {
   LOCAL_STORAGE_VULNERABILITIES_GROUPING_KEY,
@@ -21,10 +22,7 @@ import {
   VULNERABILITY_FIELDS,
 } from '../../../common/constants';
 import { useDataViewContext } from '../../../common/contexts/data_view_context';
-import {
-  LATEST_VULNERABILITIES_RETENTION_POLICY,
-  VULNERABILITIES_SEVERITY,
-} from '../../../../common/constants';
+import { VULNERABILITIES_SEVERITY } from '../../../../common/constants';
 import {
   VulnerabilitiesGroupingAggregation,
   VulnerabilitiesRootGroupingAggregation,
