+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/utils/create_detection_rule_from_vulnerability.ts
@@ -7,9 +7,9 @@
 
 import { HttpSetup } from '@kbn/core/public';
 import { i18n } from '@kbn/i18n';
-import type { Vulnerability } from '../../../../common/schemas';
+import { LATEST_VULNERABILITIES_RETENTION_POLICY } from '@kbn/cloud-security-posture-common';
+import type { Vulnerability } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import {
-  LATEST_VULNERABILITIES_RETENTION_POLICY,
   VULNERABILITIES_INDEX_PATTERN,
   VULNERABILITIES_SEVERITY,
 } from '../../../../common/constants';
