+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/utils/get_vector_score_list.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { VectorScoreBase } from '../../../../common/schemas';
+import { VectorScoreBase } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import { Vector } from '../types';
 
 export const getVectorScoreList = (vectorBaseScore: VectorScoreBase) => {
