+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/vulnerabilities_finding_flyout/vulnerability_detection_rule_counter.tsx
@@ -7,8 +7,8 @@
 
 import React from 'react';
 import type { HttpSetup } from '@kbn/core/public';
+import type { CspVulnerabilityFinding } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import { CSP_VULN_DATASET } from '../../../common/utils/get_dataset_display_name';
-import { CspVulnerabilityFinding } from '../../../../common/schemas';
 import { DetectionRuleCounter } from '../../../components/detection_rule_counter';
 import { createDetectionRuleFromVulnerabilityFinding } from '../utils/create_detection_rule_from_vulnerability';
