+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/vulnerabilities_finding_flyout/vulnerability_finding_flyout.tsx
@@ -27,13 +27,13 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { euiThemeVars } from '@kbn/ui-theme';
 import { css } from '@emotion/react';
 import { HttpSetup } from '@kbn/core-http-browser';
+import type { CspVulnerabilityFinding } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import { TakeAction } from '../../../components/take_action';
 import { truthy } from '../../../../common/utils/helpers';
 import { CspInlineDescriptionList } from '../../../components/csp_inline_description_list';
 import { VulnerabilityOverviewTab } from './vulnerability_overview_tab';
 import { VulnerabilityJsonTab } from './vulnerability_json_tab';
 import { SeverityStatusBadge } from '../../../components/vulnerability_badges';
-import type { CspVulnerabilityFinding } from '../../../../common/schemas';
 import {
   FINDINGS_VULNERABILITY_FLYOUT_DESCRIPTION_LIST,
   TAB_ID_VULNERABILITY_FLYOUT,
