+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/vulnerabilities_finding_flyout/vulnerability_json_tab.tsx
@@ -8,7 +8,7 @@
 import { CodeEditor } from '@kbn/code-editor';
 import React from 'react';
 import { XJsonLang } from '@kbn/monaco';
-import { CspVulnerabilityFinding } from '../../../../common/schemas';
+import type { CspVulnerabilityFinding } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import { JSON_TAB_VULNERABILITY_FLYOUT } from '../test_subjects';
 interface VulnerabilityJsonTabProps {
   vulnerabilityRecord: CspVulnerabilityFinding;
