+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/vulnerabilities_finding_flyout/vulnerability_overview_tab.tsx
@@ -19,8 +19,11 @@ import moment from 'moment';
 import React from 'react';
 import { euiThemeVars } from '@kbn/ui-theme';
 import { i18n } from '@kbn/i18n';
+import {
+  VectorScoreBase,
+  CspVulnerabilityFinding,
+} from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 import { getDatasetDisplayName } from '../../../common/utils/get_dataset_display_name';
-import { VectorScoreBase, CspVulnerabilityFinding } from '../../../../common/schemas';
 import { CspFlyoutMarkdown } from '../../configurations/findings_flyout/findings_flyout';
 import { NvdLogo } from '../../../assets/icons/nvd_logo_svg';
 import { CVSScoreBadge } from '../../../components/vulnerability_badges';
