+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerabilities/vulnerabilities_finding_flyout/vulnerability_table_tab.tsx
@@ -15,7 +15,7 @@ import {
 import React from 'react';
 import { getFlattenedObject } from '@kbn/std';
 import { i18n } from '@kbn/i18n';
-import { CspVulnerabilityFinding } from '../../../../common/schemas';
+import { CspVulnerabilityFinding } from '@kbn/cloud-security-posture-common/schema/vulnerabilities/latest';
 
 interface FlattenedItem {
   key: string; // flattened dot notation object path for Vulnerability;
