+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerability_dashboard/vulnerability_dashboard.test.tsx
@@ -8,10 +8,8 @@ import React from 'react';
 import Chance from 'chance';
 import { dataPluginMock } from '@kbn/data-plugin/public/mocks';
 import { unifiedSearchPluginMock } from '@kbn/unified-search-plugin/public/mocks';
-import {
-  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
-  VULN_MGMT_POLICY_TEMPLATE,
-} from '../../../common/constants';
+import { CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN } from '@kbn/cloud-security-posture-common';
+import { VULN_MGMT_POLICY_TEMPLATE } from '../../../common/constants';
 import { chartPluginMock } from '@kbn/charts-plugin/public/mocks';
 import { discoverPluginMock } from '@kbn/discover-plugin/public/mocks';
 import { useCspSetupStatusApi } from '@kbn/cloud-security-posture/src/hooks/use_csp_setup_status_api';
