+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerability_dashboard/vulnerability_table_panel_section.tsx
@@ -18,11 +18,11 @@ import {
 import { i18n } from '@kbn/i18n';
 import type { NavFilter } from '@kbn/cloud-security-posture/src/hooks/use_navigate_findings';
 import { useNavigateVulnerabilities } from '@kbn/cloud-security-posture/src/hooks/use_navigate_findings';
+import type { VulnSeverity } from '@kbn/cloud-security-posture-common';
 import {
   PatchableVulnerabilityStat,
   VulnerabilityStat,
   VulnerableResourceStat,
-  VulnSeverity,
 } from '../../../common/types_old';
 import { DASHBOARD_TABLE_TYPES } from './vulnerability_table_panel.config';
 import { VulnerabilityTablePanel } from './vulnerability_table_panel';
