+++ x-pack/plugins/cloud_security_posture/public/pages/vulnerability_dashboard/vulnerability_trend_graph.tsx
@@ -20,8 +20,9 @@ import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
 
 import { useNavigateVulnerabilities } from '@kbn/cloud-security-posture/src/hooks/use_navigate_findings';
+import type { VulnSeverity } from '@kbn/cloud-security-posture-common';
 import { truthy } from '../../../common/utils/helpers';
-import { VulnStatsTrend, VulnSeverity } from '../../../common/types_old';
+import { VulnStatsTrend } from '../../../common/types_old';
 import { useVulnerabilityDashboardApi } from '../../common/api/use_vulnerability_dashboard_api';
 import { getSeverityStatusColor } from '../../common/utils/get_vulnerability_colors';
 import { ChartPanel } from '../../components/chart_panel';
