+++ x-pack/plugins/cloud_security_posture/server/create_indices/latest_indices.ts
@@ -5,13 +5,15 @@
  * 2.0.
  */
 
-import { CDR_LATEST_NATIVE_MISCONFIGURATIONS_INDEX_PATTERN } from '@kbn/cloud-security-posture-common';
+import {
+  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
+  CDR_LATEST_NATIVE_MISCONFIGURATIONS_INDEX_PATTERN,
+} from '@kbn/cloud-security-posture-common';
 import {
   FINDINGS_INDEX_NAME,
   LATEST_FINDINGS_INDEX_TEMPLATE_NAME,
   LATEST_FINDINGS_INDEX_DEFAULT_NS,
   VULNERABILITIES_INDEX_NAME,
-  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
   LATEST_VULNERABILITIES_INDEX_TEMPLATE_NAME,
 } from '../../common/constants';
 import { LatestIndexConfig } from './types';
