+++ x-pack/plugins/cloud_security_posture/server/routes/status/status.ts
@@ -12,6 +12,9 @@ import {
   STATUS_ROUTE_PATH,
   LATEST_FINDINGS_RETENTION_POLICY,
   CDR_MISCONFIGURATIONS_INDEX_PATTERN,
+  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
+  LATEST_VULNERABILITIES_RETENTION_POLICY,
+  CDR_VULNERABILITIES_INDEX_PATTERN,
 } from '@kbn/cloud-security-posture-common';
 import type {
   CspSetupStatus,
@@ -36,11 +39,8 @@ import {
   BENCHMARK_SCORE_INDEX_DEFAULT_NS,
   VULNERABILITIES_INDEX_PATTERN,
   POSTURE_TYPES,
-  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
   VULN_MGMT_POLICY_TEMPLATE,
   POSTURE_TYPE_ALL,
-  LATEST_VULNERABILITIES_RETENTION_POLICY,
-  CDR_VULNERABILITIES_INDEX_PATTERN,
 } from '../../../common/constants';
 import type {
   CspApiRequestHandlerContext,
