+++ x-pack/plugins/cloud_security_posture/server/saved_objects/data_views.ts
@@ -19,11 +19,11 @@ import {
   CDR_MISCONFIGURATIONS_INDEX_PATTERN,
   CDR_MISCONFIGURATIONS_DATA_VIEW_ID_PREFIX,
   CDR_MISCONFIGURATIONS_DATA_VIEW_NAME,
+  CDR_VULNERABILITIES_INDEX_PATTERN,
 } from '@kbn/cloud-security-posture-common';
 import {
   CDR_VULNERABILITIES_DATA_VIEW_ID_PREFIX,
   CDR_VULNERABILITIES_DATA_VIEW_NAME,
-  CDR_VULNERABILITIES_INDEX_PATTERN,
 } from '../../common/constants';
 
 const DATA_VIEW_TIME_FIELD = '@timestamp';
