+++ x-pack/plugins/security_solution_serverless/server/cloud_security/constants.ts
@@ -9,11 +9,9 @@ import {
   CSPM_POLICY_TEMPLATE,
   KSPM_POLICY_TEMPLATE,
   CDR_LATEST_NATIVE_MISCONFIGURATIONS_INDEX_PATTERN,
-} from '@kbn/cloud-security-posture-common';
-import {
-  CNVM_POLICY_TEMPLATE,
   CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
-} from '@kbn/cloud-security-posture-plugin/common/constants';
+} from '@kbn/cloud-security-posture-common';
+import { CNVM_POLICY_TEMPLATE } from '@kbn/cloud-security-posture-plugin/common/constants';
 import { INTEGRATION_PACKAGE_NAME } from '@kbn/cloud-defend-plugin/common/constants';
 
 export const CLOUD_DEFEND_HEARTBEAT_INDEX = 'metrics-cloud_defend.heartbeat-*';
