+++ x-pack/test/cloud_security_posture_api/routes/helper/user_roles_utilites.ts
@@ -5,10 +5,12 @@
  * 2.0.
  */
 
-import { CDR_LATEST_NATIVE_MISCONFIGURATIONS_INDEX_PATTERN } from '@kbn/cloud-security-posture-common';
+import {
+  CDR_LATEST_NATIVE_MISCONFIGURATIONS_INDEX_PATTERN,
+  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
+} from '@kbn/cloud-security-posture-common';
 import {
   BENCHMARK_SCORE_INDEX_PATTERN,
-  CDR_LATEST_NATIVE_VULNERABILITIES_INDEX_PATTERN,
   ALERTS_INDEX_PATTERN,
   FINDINGS_INDEX_PATTERN,
 } from '@kbn/cloud-security-posture-plugin/common/constants';
